Nearly there

+++ cuda_rasterizer/forward.h
@@ -86,7 +86,7 @@ namespace FORWARD
 		int width,
 		int height,
 		const uint32_t *splat_ids,
-		const ushort2 *splat_id_ranges,
+		const uint2 *splat_id_ranges,
 		const float2 *means_2d,
 		const float4 *conic_opacities,
 		const float *depths,
